% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_brick.R
\name{build_brick}
\alias{build_brick}
\title{Builds a raster brick from a netCDF list}
\usage{
build_brick(x, y, z)
}
\arguments{
\item{x}{The length of the raster brick in its x dimension. Can be read from the netCDF info.}

\item{y}{The length of the raster brick in its y dimension. Can be read from the netCDF info.}

\item{z}{The length of the brick in its time dimension (number of brick layers).}
}
\value{

}
\description{
Builds a raster brick from a netCDF list
}
