% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncdf.R
\name{load_ncdf}
\alias{load_ncdf}
\title{SPI spatial calculator}
\usage{
load_ncdf(path, filename)
}
\arguments{
\item{path}{The local path to where the nedCDF file is located}

\item{filename}{The filename of the netCDF. Must include the ending ".nc"}
}
\value{
Returns a netCDF object and prints its respective metadata
}
\description{
SPI spatial calculator
}
