% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick.R
\name{ncdf_brick}
\alias{ncdf_brick}
\title{SPI spatial calculator}
\usage{
ncdf_brick(path, filename, x, y, z)
}
\arguments{
\item{path}{The local path to where the nedCDF file is located}

\item{filename}{The filename of the netCDF. Must include the ending ".nc"}

\item{x}{The length of the raster brick in its x dimension. Can be read from the netCDF info.}

\item{y}{The length of the raster brick in its y dimension. Can be read from the netCDF info.}

\item{z}{The length of the brick in its time dimension (number of brick layers).}
}
\value{
Returns a raster brick
}
\description{
SPI spatial calculator
}
