% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIspatial.R
\name{SPIspatial}
\alias{SPIspatial}
\title{Calculates SPI for every layer of raster brick}
\usage{
SPIspatial(brick, timescale, start, frequency, ref.start, ref.end)
}
\arguments{
\item{brick}{Output brick from the build_brick function}

\item{timescale}{From SPEI::spi. an integer, representing the time scale at which the SPEI / SPI will be computed.}

\item{start}{When was the first measurement in the dataset? A vector with two numbers: e.g. c(1970, 1)}

\item{frequency}{integer. If the timespan covered in the dataset are month, then use 12.}

\item{ref.start}{When does the reference period start? A vector of two numbers: e.g. c (1980, 1)}

\item{ref.end}{When does the reference period end? A vector of two numbers: e.g. c (2010, 1)}
}
\value{
A raster brick
}
\description{
Calculates SPI for every layer of raster brick
}
